# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Alpogo API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/item/type": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "A\u00f1ade un tipo de item",
                "description": "",
                "operationId": "add_item_type",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ItemType"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/item/type/{slug}": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "Detalle de un tipo de item",
                "description": "",
                "operationId": "get_item_type",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "410": {
                        "description": "Item Type not found"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "item"
                ],
                "summary": "Edita un tipo de item",
                "description": "",
                "operationId": "update_item_type",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ItemType"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "410": {
                        "description": "Item Type not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/item/type/{id}": {
            "delete": {
                "tags": [
                    "item"
                ],
                "summary": "Elimina un tipo de item",
                "description": "",
                "operationId": "destroy_item_type",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "410": {
                        "description": "Item type not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/event/{slug}/promotion": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "A\u00f1ade una promoci\u00f3n al evento",
                "description": "",
                "operationId": "get_promotion_from_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "A\u00f1ade una promoci\u00f3n al evento",
                "description": "",
                "operationId": "add_promotion_to_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Promotion"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/event/{slug}/promotion/{id}": {
            "put": {
                "tags": [
                    "event"
                ],
                "summary": "Edita una promoci\u00f3n de un evento",
                "description": "",
                "operationId": "update_promotion_of_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Promotion"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "409": {
                        "description": "Ticket not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "event"
                ],
                "summary": "Edita parcialmente una promoci\u00f3n de un evento",
                "description": "",
                "operationId": "partial_update_promotion_of_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Promotion"
                        }
                    }
                ],
                "responses": {
                    "408": {
                        "description": "Event not found"
                    },
                    "409": {
                        "description": "Ticket not found"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ItemType": {
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "jhondoe"
                }
            },
            "type": "object",
            "xml": {
                "name": "ItemType"
            }
        },
        "Promotion": {
            "required": [
                "detail",
                "quantity",
                "amount",
                "producer_amount"
            ],
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "El detalle de Jhon"
                },
                "quantity": {
                    "type": "string",
                    "example": "21"
                },
                "amount": {
                    "type": "string",
                    "example": "123"
                },
                "producer_amount": {
                    "type": "string",
                    "example": "1234"
                }
            },
            "type": "object",
            "xml": {
                "name": "Promotion"
            }
        },
...
    }
}